requirements.txt
# Core dependencies for Stock Predictor with LLMs
# Install with: pip install -r requirements.txt

# Environment and configuration
python-dotenv==1.2.1

# News API
newsapi-python==0.2.7

# HTTP requests
requests==2.32.5

# Data processing
pandas>=2.0.0

# Stock data fetching
yfinance>=0.2.0

